requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
chardet==4.0.0
colorama==0.4.4
discord.ext.context==0.1.7
discord.py==1.7.2
idna==3.2
multidict==5.1.0
termcolor==1.1.0
typing-extensions==3.10.0.0
yarl==1.6.3
